Adding the crow sprite and some tools to build palette lookups and all that shit. my brain hurts. but at least I have the sick intro done besides the music. higher framerate than anyone else has had... whatever.

+++ L1DemoJamisNemo.X/drawing_helpers.h
@@ -19,7 +19,7 @@ extern "C" {
 float radians(uint16_t angle);
 int realtoint(float oldval, float oldmin, float oldmax, float newmin, float newmax);
 
-void rcc_color(unsigned int color);
+void rcc_color(uint16_t color);
 void rcc_setdest(__eds__ uint8_t *buf);
 void rcc_draw(uint16_t x, uint16_t y, uint16_t w, uint16_t h);
 void rcc_w1tow2(__eds__ uint8_t *dest, __eds__ uint8_t *src);
